fix: ProcessingJobResponse progress 타입 float → int 수정 (이레님 리뷰 반영)

+++ app/dtos/processing_jobs.py
@@ -18,5 +18,5 @@ class ProcessingJobResponse(BaseModel):
     job_type: str
     status: str
     record_id: int | None = None
-    progress: float | None = None
+    progress: int | None = None
     result_ref: str | None = None
